+++ samples/boards/st/power_mgmt/stop3/src/main.c
@@ -10,7 +10,7 @@
 #include <zephyr/drivers/gpio.h>
 #include <zephyr/sys/printk.h>
 #include <stm32_ll_pwr.h>
-#include <gpio/gpio_stm32.h>
+#include <stm32_gpio_shared.h>
 
 #define SLEEP_TIME_MS   2000
